+++ packages/discovery-react-components/src/components/CIDocument/components/CIDocument/CIDocument.tsx
@@ -34,7 +34,16 @@ import {
 } from 'utils/document/nonContractUtils';
 import { withErrorBoundary, WithErrorBoundaryProps } from 'utils/hoc/withErrorBoundary';
 import { Filter, FilterGroup, FilterChangeArgs } from '../FilterPanel/types';
-import { MetadataData, EnrichedHtml, Contract, Item, Field } from 'components/CIDocument/types';
+import {
+  Metadata,
+  MetadataData,
+  EnrichedHtml,
+  Contract,
+  Item,
+  Field,
+  Attributes,
+  Relations
+} from 'components/CIDocument/types';
 import { Address, Mention } from '../MetadataPane/types';
 import { Items } from '../DetailsPane/types';
 import { defaultTheme, Theme } from 'utils/theme';
@@ -81,6 +90,9 @@ interface State {
     byItem: any;
     bySection: any;
   };
+  metadata: Metadata[];
+  attributes: Attributes[];
+  relations: Relations[];
 }
 
 enum ActionType {
@@ -98,7 +110,10 @@ const INITIAL_STATE: State = {
   isError: false,
   styles: [],
   sections: [],
-  itemMap: { byItem: {}, bySection: {} }
+  itemMap: { byItem: {}, bySection: {} },
+  metadata: [],
+  attributes: [],
+  relations: []
 };
 
 const RELATIONS = 'relations';
@@ -160,11 +175,11 @@ const CIDocument: FC<CIDocumentProps> = ({
   const [selectedContractFilter, setSelectedContractFilter] = useState(FILTERS);
 
   const enrichment = get(enrichedHtml, enrichmentName, {});
-  const { elements = [], metadata = [], parties = [] }: Contract = enrichment;
+  const { elements = [], parties = [] }: Contract = enrichment;
 
   let itemList = elements;
   if (isInvoiceOrPurchaseOrder(enrichmentName)) {
-    itemList = get(enrichment, selectedType, []);
+    itemList = state[selectedType] || [];
   }
 
   const resetTabs = (): void => {
@@ -189,7 +204,14 @@ const CIDocument: FC<CIDocumentProps> = ({
           const doc = await processDoc(document, { sections: true, itemMap: true });
           dispatch({
             type: ActionType.SET,
-            data: { styles: doc.styles, sections: doc.sections, itemMap: doc.itemMap }
+            data: {
+              styles: doc.styles,
+              sections: doc.sections,
+              itemMap: doc.itemMap,
+              metadata: doc.metadata,
+              attributes: doc.attributes,
+              relations: doc.relations
+            }
           });
         } catch (err) {
           dispatch({ type: ActionType.ERROR, data: err });
@@ -342,7 +364,7 @@ const CIDocument: FC<CIDocumentProps> = ({
             <Tab tabIndex={0} label={messages.metadataTabLabel}>
               {!hasError && (
                 <MetadataPane
-                  metadata={metadata}
+                  metadata={state.metadata}
                   activeMetadataId={activeMetadataIds[0]}
                   parties={parties}
                   messages={messages}