+++ packages/discovery-react-components/src/components/StructuredQuery/__tests__/StructuredQuery.test.tsx
@@ -131,8 +131,8 @@ describe('<StructuredQuery />', () => {
         expect(values.length).toEqual(1);
       });
 
-      test('and the one top-level rule row does not have a Remove row button', () => {
-        const removeRuleButton = structuredQuery.queryByLabelText('Remove row');
+      test('and the one top-level rule row does not have a Remove rule button', () => {
+        const removeRuleButton = structuredQuery.queryByTestId('remove-rule-row-button-0');
         expect(removeRuleButton).toBe(null);
       });
 
@@ -158,7 +158,7 @@ describe('<StructuredQuery />', () => {
       });
 
       test('and both rule rows now include the Remove rule button', () => {
-        const removeRuleButtons = structuredQuery.queryAllByLabelText('Remove rule');
+        const removeRuleButtons = structuredQuery.queryAllByTestId('remove-rule-row-button-0');
         expect(removeRuleButtons.length).toEqual(2);
       });
     });
@@ -167,7 +167,7 @@ describe('<StructuredQuery />', () => {
       beforeEach(() => {
         const addRuleButton = structuredQuery.getByText('Add rule');
         addRuleButton.click();
-        const removeRuleButton = structuredQuery.getAllByTestId('remove-rule-row-button')[1];
+        const removeRuleButton = structuredQuery.queryAllByTestId('remove-rule-row-button-0')[1];
         removeRuleButton.click();
       });
 
@@ -181,7 +181,7 @@ describe('<StructuredQuery />', () => {
       });
 
       test('and no Remove rule buttons are displayed since only one rule row still remains', () => {
-        const removeRuleButton = structuredQuery.queryByLabelText('Remove row');
+        const removeRuleButton = structuredQuery.queryByTestId('remove-rule-row-button-0');
         expect(removeRuleButton).toBe(null);
       });
     });
@@ -200,6 +200,191 @@ describe('<StructuredQuery />', () => {
         addRuleButton.click();
         const getAddRuleButtonAgain = structuredQuery.queryByText('Add rule');
         expect(getAddRuleButtonAgain).toBe(null);
+      });
+    });
+  });
+
+  describe('nested rule groups', () => {
+    let structuredQuery: RenderResult;
+    beforeEach(() => {
+      structuredQuery = render(<StructuredQuery />);
+    });
+
+    describe('adding new nested rule groups', () => {
+      describe('when the Add group of rules button is initially clicked', () => {
+        beforeEach(() => {
+          const addRuleGroupButton = structuredQuery.getByText('Add group of rules');
+          addRuleGroupButton.click();
+        });
+
+        test('one new nested rule group is added', () => {
+          const ruleGroups = structuredQuery.queryAllByTestId('rule-group');
+          expect(ruleGroups.length).toEqual(2);
+        });
+
+        test('and it has a Remove rule button', () => {
+          const removeRuleRowButton = structuredQuery.queryAllByTestId('remove-rule-row-button-1');
+          expect(removeRuleRowButton.length).toEqual(1);
+        });
+
+        test('and it has a button to add a rule', () => {
+          const addRuleRowButtons = structuredQuery.queryAllByText('Add rule');
+          expect(addRuleRowButtons.length).toEqual(2);
+        });
+
+        test('and it does not have a button to add another level of rule groups', () => {
+          const addRuleGroupButtons = structuredQuery.queryAllByText('Add group of rules');
+          expect(addRuleGroupButtons.length).toEqual(1);
+        });
+      });
+
+      describe('when adding multiple groups of rules', () => {
+        let addRuleGroupButton: HTMLElement;
+        beforeEach(() => {
+          addRuleGroupButton = structuredQuery.getByText('Add group of rules');
+          addRuleGroupButton.click();
+          addRuleGroupButton.click();
+        });
+
+        test('a new group of rules is added for each click of the Add group of rules button, with one rule row each', () => {
+          const ruleGroups = structuredQuery.queryAllByTestId('rule-group');
+          const ruleRowsTopLevel = structuredQuery.queryAllByTestId('rule-row-0');
+          const ruleRowsGroupZero = structuredQuery.queryAllByTestId('rule-row-1');
+          const ruleRowsGroupOne = structuredQuery.queryAllByTestId('rule-row-2');
+          expect(ruleGroups.length).toEqual(3);
+          expect(ruleRowsTopLevel.length).toEqual(1);
+          expect(ruleRowsGroupZero.length).toEqual(1);
+          expect(ruleRowsGroupOne.length).toEqual(1);
+        });
+
+        test('the Add group of rule buttons is no longer displayed when the max number of nested rule groups has been added', () => {
+          addRuleGroupButton.click();
+          addRuleGroupButton.click();
+          addRuleGroupButton = structuredQuery.getByText('Add group of rules');
+          expect(addRuleGroupButton).toBeDefined();
+          addRuleGroupButton.click();
+          const addRuleGroupButtonQuery = structuredQuery.queryByText('Add group of rules');
+          expect(addRuleGroupButtonQuery).toBe(null);
+        });
+      });
+    });
+
+    describe('adding rule rows to nested rule groups', () => {
+      let addRuleGroupButton: HTMLElement;
+      let addRuleToGroupZeroButton: HTMLElement;
+      beforeEach(() => {
+        addRuleGroupButton = structuredQuery.getByText('Add group of rules');
+        addRuleGroupButton.click();
+        addRuleToGroupZeroButton = structuredQuery.queryAllByText('Add rule')[0];
+        addRuleToGroupZeroButton.click();
+      });
+      test('each click of the Add rule button adds one new rule to the correct nested rule group', () => {
+        let ruleRowsGroupZero = structuredQuery.queryAllByTestId('rule-row-1');
+        expect(ruleRowsGroupZero.length).toEqual(2);
+        addRuleGroupButton.click();
+        const addRuleRowToGroupOneButton = structuredQuery.queryAllByText('Add rule')[1];
+        addRuleRowToGroupOneButton.click();
+        let ruleRowsGroupOne = structuredQuery.queryAllByTestId('rule-row-2');
+        expect(ruleRowsGroupOne.length).toEqual(2);
+        addRuleToGroupZeroButton.click();
+        ruleRowsGroupZero = structuredQuery.queryAllByTestId('rule-row-1');
+        expect(ruleRowsGroupZero.length).toEqual(3);
+        addRuleRowToGroupOneButton.click();
+        ruleRowsGroupOne = structuredQuery.queryAllByTestId('rule-row-2');
+        expect(ruleRowsGroupOne.length).toEqual(3);
+      });
+      test('no more rules can be added to a nested rule group past the max number of allowed rule row siblings', () => {
+        expect(structuredQuery.queryAllByText('Add rule').length).toBe(2);
+        addRuleToGroupZeroButton.click();
+        expect(structuredQuery.queryAllByText('Add rule').length).toBe(1);
+      });
+    });
+
+    describe('removing rule rows from nested rule groups', () => {
+      let addRuleGroupButton: HTMLElement;
+      beforeEach(() => {
+        addRuleGroupButton = structuredQuery.getByText('Add group of rules');
+        addRuleGroupButton.click();
+      });
+      describe('when there are multiple nested rule row groups', () => {
+        let addRuleRowToGroupOneButton: HTMLElement;
+        let addRuleRowToGroupTwoButton: HTMLElement;
+        // This sets up two nested rule groups, with two rule rows in group one and three rule rows in group two
+        beforeEach(() => {
+          addRuleGroupButton.click();
+          addRuleGroupButton.click();
+          addRuleRowToGroupOneButton = structuredQuery.queryAllByText('Add rule')[1];
+          addRuleRowToGroupOneButton.click();
+          addRuleRowToGroupTwoButton = structuredQuery.queryAllByText('Add rule')[2];
+          addRuleRowToGroupTwoButton.click();
+          addRuleRowToGroupTwoButton.click();
+        });
+        test('clicking Remove rule buttons in group one removes rules from group one', () => {
+          let ruleRowsGroupOne = structuredQuery.queryAllByTestId('rule-row-2');
+          expect(ruleRowsGroupOne.length).toEqual(2);
+          const removeRuleRowFromGroupOneButton = structuredQuery.queryAllByTestId(
+            'remove-rule-row-button-2'
+          )[0];
+          removeRuleRowFromGroupOneButton.click();
+          ruleRowsGroupOne = structuredQuery.queryAllByTestId('rule-row-2');
+          expect(ruleRowsGroupOne.length).toEqual(1);
+        });
+        test('clicking Remove rule buttons in group two removes rules from group two', () => {
+          let ruleRowsGroupTwo = structuredQuery.queryAllByTestId('rule-row-3');
+          expect(ruleRowsGroupTwo.length).toEqual(3);
+          let removeRuleRowFromGroupTwoButton = structuredQuery.queryAllByTestId(
+            'remove-rule-row-button-3'
+          )[0];
+          removeRuleRowFromGroupTwoButton.click();
+          ruleRowsGroupTwo = structuredQuery.queryAllByTestId('rule-row-3');
+          expect(ruleRowsGroupTwo.length).toEqual(2);
+          removeRuleRowFromGroupTwoButton = structuredQuery.queryAllByTestId(
+            'remove-rule-row-button-3'
+          )[0];
+          removeRuleRowFromGroupTwoButton.click();
+          ruleRowsGroupTwo = structuredQuery.queryAllByTestId('rule-row-3');
+          expect(ruleRowsGroupTwo.length).toEqual(1);
+          removeRuleRowFromGroupTwoButton = structuredQuery.queryAllByTestId(
+            'remove-rule-row-button-3'
+          )[0];
+          removeRuleRowFromGroupTwoButton.click();
+          const ruleRowsGroupTwoQuery = structuredQuery.queryByTestId('rule-row-3');
+          expect(ruleRowsGroupTwoQuery).toBe(null);
+        });
+      });
+      describe('when there is a singular nested rule row group', () => {
+        describe('when the Remove rule icon is clicked and it is not the last remaining rule row in a nested rule group', () => {
+          test('one click removes one rule row from the correct nested rule group', () => {
+            const addRuleButton = structuredQuery.queryAllByText('Add rule')[0];
+            addRuleButton.click();
+            addRuleButton.click();
+            const removeRuleRowButtons = structuredQuery.queryAllByText('Remove rule');
+            expect(structuredQuery.queryAllByTestId('rule-row-1').length).toEqual(3);
+            removeRuleRowButtons[2].click();
+            expect(structuredQuery.queryAllByTestId('rule-row-1').length).toEqual(2);
+            removeRuleRowButtons[1].click();
+            expect(structuredQuery.queryAllByTestId('rule-row-1').length).toEqual(1);
+          });
+        });
+        describe('when the Remove rule icon is clicked for the last remaining rule row in a nested rule group', () => {
+          test('the entire nested group is removed', () => {
+            const addRuleButton = structuredQuery.queryAllByText('Add rule')[0];
+            addRuleButton.click();
+            addRuleButton.click();
+            const removeRuleRowButtons = structuredQuery.queryAllByText('Remove rule');
+            removeRuleRowButtons[2].click();
+            removeRuleRowButtons[1].click();
+            let ruleGroups = structuredQuery.queryAllByTestId('rule-group');
+            let nestedRuleGroupZero = structuredQuery.queryByTestId('rule-row-1');
+            expect(ruleGroups.length).toEqual(2);
+            expect(nestedRuleGroupZero).toBeDefined();
+            removeRuleRowButtons[0].click();
+            ruleGroups = structuredQuery.queryAllByTestId('rule-group');
+            nestedRuleGroupZero = structuredQuery.queryByTestId('rule-row-1');
+            expect(ruleGroups.length).toEqual(1);
+            expect(nestedRuleGroupZero).toBe(null);
+          });
+        });
       });
     });
   });